fraud-service/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic-settings>=2.0
jinja2>=3.1
joblib>=1.3
numpy>=1.26
httpx>=0.27
pytest>=8.0
